+++ src/main/groovy/com/cedarsoftware/ncube/UrlCommandCell.groovy
@@ -32,7 +32,6 @@ import static com.cedarsoftware.ncube.NCubeAppContext.ncubeRuntime
 @CompileStatic
 abstract class UrlCommandCell implements CommandCell
 {
-    protected static NCubeRuntimeClient ncubeRuntimeClient
     private static final Logger LOG = LoggerFactory.getLogger(UrlCommandCell.class)
     private String cmd
     private volatile transient String errorMsg = null
@@ -79,44 +78,27 @@ abstract class UrlCommandCell implements CommandCell
         return cacheable
     }
 
-//    void clearClassLoaderCache()
-//    {
-//        // classpath case, lets clear all classes before setting to null.
-//        def localVar = cache
-//        if (localVar instanceof GroovyClassLoader)
-//        {
-//            (localVar as GroovyClassLoader).clearCache()
-//        }
-//        cache = null
-//    }
-
-    // When no L3, use this (also see GroovyBase)
     void clearClassLoaderCache()
     {
+        TimedSynchronize.synchronize(hasBeenCachedLock, 200, TimeUnit.MILLISECONDS, 'Dead lock detected attempting to clear ClassLoader cache.')
         hasBeenCached = false
-        if (cache == null)
-        {
-            return
-        }
+        def localVar = cache
 
-        synchronized (lock)
+        try
         {
-            if (cache == null)
-            {
-                return
-            }
-
             // classpath case, lets clear all classes before setting to null.
-            if (cache instanceof GroovyClassLoader)
+            if (localVar instanceof GroovyClassLoader)
             {
-                ((GroovyClassLoader)cache).clearCache()
+                ((GroovyClassLoader)localVar).clearCache()
             }
             cache = null
         }
+        finally
+        {
+            hasBeenCachedLock.unlock();
+        }
     }
-
-    protected Object getLock() { return '' }
-
+    
     protected URL getActualUrl(Map<String, Object> ctx)
     {
         for (int i=0; i < 2; i++)
@@ -128,15 +110,15 @@ abstract class UrlCommandCell implements CommandCell
             catch(Exception e)
             {
                 NCube cube = getNCube(ctx)
+                String where = "url: ${getUrl()}, cube: ${cube.name}, app: ${cube.applicationID}"
                 if (i == 1)
                 {   // Note: Error is marked, it will not be retried in the future
-                    errorMessage = "Invalid URL in cell (malformed or cannot resolve given classpath): " + getUrl() + ", cube: " + cube.name + ", app: " + cube.applicationID
-                    LOG.warn(getClass().simpleName + ': failed 2nd attempt [will NOT retry in future] getActualUrl() - unable to resolve against sys.classpath, url: ' + getUrl() + ", cube: " + cube.name + ", app: " + cube.applicationID)
-                    throw new IllegalStateException(errorMessage, e)
+                    LOG.warn("${getClass().simpleName}: failed 2nd attempt [will NOT retry in future] getActualUrl() - unable to resolve against sys.classpath, ${where}")
+                    throw new IllegalStateException("Invalid URL in cell (unable to resolve against sys.classpath), ${where}", e)
                 }
                 else
                 {
-                    LOG.warn(getClass().simpleName + ': retrying getActualUrl() - unable to resolve against sys.classpath, url: ' + getUrl() + ", cube: " + cube.name + ", app: " + cube.applicationID)
+                    LOG.warn("${getClass().simpleName}: retrying getActualUrl() - unable to resolve against sys.classpath, ${where}")
                     Thread.sleep(100)
                 }
             }