+++ src/test/resources/static/tests/ncube/cp1/factorial.groovy
@@ -1,0 +1,2 @@
+def fac = { n -> n == 0 ? 1 : n * call(n - 1) }
+fac(input.x)